Images: Extracted out image resizing to its own class
